--- common/test_node/CMakeLists.txt
@@ -1,32 +1,0 @@
-cmake_minimum_required(VERSION 3.8)
-project(test_node)
-
-find_package(ament_cmake_auto REQUIRED)
-find_package(std_msgs REQUIRED)
-ament_auto_find_build_dependencies()
-
-autoware_package()
-
-include_directories(
-  include)
-
-set(executable_name ${PROJECT_NAME})
-set(library_name ${executable_name}_core)
-
-ament_auto_add_library(${library_name} SHARED
-  src/test_node.cpp)
-ament_target_dependencies(${library_name} rclcpp std_msgs)
-
-rclcpp_components_register_node(${library_name}
-  PLUGIN "test_node::TestNode"
-  EXECUTABLE ${executable_name})
-
-ament_auto_add_executable(test_pub src/test_pub.cpp)
-ament_target_dependencies(test_pub rclcpp std_msgs)
-
-install(DIRECTORY
-  launch
-  DESTINATION share/${PROJECT_NAME}/
-)
-
-ament_auto_package()
