--- common/test_node/include/test_node/test_node.hpp
@@ -1,40 +1,0 @@
-// Copyright 2024 The Autoware Contributors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#ifndef TEST_NODE__TEST_NODE_HPP_
-#define TEST_NODE__TEST_NODE_HPP_
-
-#include "autoware_node/autoware_node.hpp"
-#include "rclcpp_lifecycle/lifecycle_node.hpp"
-#include "test_node/visibility_control.hpp"
-
-#include "std_msgs/msg/string.hpp"
-
-namespace test_node
-{
-
-class TestNode : public autoware_node::AutowareNode
-{
-public:
-  TEST_NODE_PUBLIC
-  explicit TestNode(const rclcpp::NodeOptions & options = rclcpp::NodeOptions());
-
-private:
-  void topic_callback(const std_msgs::msg::String::SharedPtr msg);
-  rclcpp::Subscription<std_msgs::msg::String>::SharedPtr monitored_subscription_;
-};
-
-}  // namespace test_node
-
-#endif  // TEST_NODE__TEST_NODE_HPP_
